chore: refactor to move endpageblock to a generic folder
